Removed phase angle in transformation and swap qmk4 and umk4
We are now using the Dec 12, 2016 version of Alfred's calibration. This fixes the bugs in
the previous versions that required the need for a phase angle in the coordinate
transformation from cartesian to tangential polarization. The new calibration also swaps
q and u polarization. I swapped 'qmk4' and 'umk4' in the coordinate transformation to accommodate
the change in Alfred's code.