Only unlock state if locked it in the first place
There was a bug that the state was getting unlocked no matter what each time kcor_rt was called. So if a set of files was large enough to take more than two cron intervals, the kcor_rt call after the first interval would unlock the state, and the call after the second one would start running a second process at the same time as the previous one, causing weird errors.